requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
streamlit
twilio
fitbit
google-generative-ai
